use long double for sqrt instead of regular double

+++ src/ppmw.c
@@ -12,7 +12,7 @@
 void ppmw_proc(long start, long end, int rank, int size) {
 	//discover primes up to the specified number.
 	for (long num = start; num <= end; num++) {
-		long numToCheck = (long)sqrt(num) + 1;
+		long numToCheck = (long)sqrt((long double)num) + 1;
 		//basic preliminary checks to determine if we can simply skip this num.
 		if (!pp_prelim_check(num)) {
 			MPI_Barrier(MPI_COMM_WORLD);
